+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - Added help messages for commands:
     - `odoo-helper install pre-requirements`
     - `odoo-helper install sys-deps`
+    - `odoo-helper install py-deps`
     - `odoo-helper install reinstall-odoo`
     - `odoo-helper install reinstall-venv`
 - New options to `odoo-install` command